+++ src/getStepStartStates.ts
@@ -30,7 +30,10 @@ import {
 import { steps as postgreSqlDatabaseSteps } from './steps/resource-manager/databases/postgresql/constants';
 import { steps as sqlDatabaseSteps } from './steps/resource-manager/databases/sql/constants';
 import { STEP_RM_COMPUTE_NETWORK_RELATIONSHIPS } from './steps/resource-manager/interservice/constants';
-import { STEP_RM_KEYVAULT_VAULTS } from './steps/resource-manager/key-vault/constants';
+import {
+  STEP_RM_KEYVAULT_VAULTS,
+  KeyVaultStepIds,
+} from './steps/resource-manager/key-vault/constants';
 import {
   STEP_RM_NETWORK_INTERFACES,
   STEP_RM_NETWORK_LOAD_BALANCERS,
@@ -143,6 +146,7 @@ export function getResourceManagerSteps(): GetApiSteps {
   return {
     executeFirstSteps: [
       STEP_RM_KEYVAULT_VAULTS,
+      KeyVaultStepIds.KEY_VAULT_PRINCIPAL_RELATIONSHIPS,
       STEP_RM_NETWORK_VIRTUAL_NETWORKS,
       STEP_RM_NETWORK_SECURITY_GROUPS,
       STEP_RM_NETWORK_SECURITY_GROUP_RULE_RELATIONSHIPS,