+++ src/steps/resource-manager/constants.ts
@@ -3,8 +3,15 @@
  * It includes non-resourced based Azure scopes, such as Azure Subscriptions and Azure Resource Groups.
  */
 export const ANY_SCOPE = 'ANY_SCOPE';
+
 /**
  * ANY_RESOURCE is used to describe any Azure resource.
  * It does NOT include non-resourced based Azure scopes, such as Azure Subscriptions and Azure Resource Groups.
  */
 export const ANY_RESOURCE = 'ANY_RESOURCE';
+
+/**
+ * ANY_PRINCIPAL is used to describe any active directory principal.
+ * It includes users, groups, and service principals.
+ */
+export const ANY_PRINCIPAL = 'ANY_PRINCIPAL';